requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
email-validator
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
itsdangerous
openai
jinja2
python-multipart
boto3
Pillow
img2pdf
stripe
PyPDF2
reportlab
passlib
requests
